+++ react/ProductSearchContextProvider.js
@@ -4,11 +4,56 @@ import { Query } from 'react-apollo'
 import { Helmet, withRuntimeContext } from 'render'
 
 import DataLayerApolloWrapper from './components/DataLayerApolloWrapper'
-import { processSearchContextProps } from './helpers/searchHelpers'
 import searchQuery from './queries/searchQuery.gql'
 
 const DEFAULT_PAGE = 1
 const DEFAULT_MAX_ITEMS_PER_PAGE = 1
+
+const SORT_OPTIONS = [
+  {
+    value: 'OrderByTopSaleDESC',
+    label: 'ordenation.sales',
+  },
+  {
+    value: 'OrderByReleaseDateDESC',
+    label: 'ordenation.release.date',
+  },
+  {
+    value: 'OrderByBestDiscountDESC',
+    label: 'ordenation.discount',
+  },
+  {
+    value: 'OrderByPriceDESC',
+    label: 'ordenation.price.descending',
+  },
+  {
+    value: 'OrderByPriceASC',
+    label: 'ordenation.price.ascending',
+  },
+  {
+    value: 'OrderByNameASC',
+    label: 'ordenation.name.ascending',
+  },
+  {
+    value: 'OrderByNameDESC',
+    label: 'ordenation.name.descending',
+  },
+]
+
+function createMap(params, rest, isBrand) {
+  const paramValues = Object.keys(params)
+    .filter(param => !param.startsWith('_'))
+    .map(key => params[key])
+    .concat(rest ? rest.split(',') : [])
+  const map =
+    paramValues.length > 0 &&
+    Array(paramValues.length - 1)
+      .fill('c')
+      .join(',') +
+    (paramValues.length > 1 ? ',' : '') +
+    (isBrand ? 'b' : 'c')
+  return map
+}
 
 const canonicalPathFromParams = ({
   brand,
@@ -31,6 +76,13 @@ class ProductSearchContextProvider extends Component {
     runtime: PropTypes.shape({
       page: PropTypes.string.isRequired,
     }),
+    query: PropTypes.shape({
+      map: PropTypes.string,
+      rest: PropTypes.string,
+      order: PropTypes.oneOf(SORT_OPTIONS.map(o => o.value)),
+      priceRange: PropTypes.string,
+    }),
+    nextTreePath: PropTypes.string,
   }
 
   state = {
@@ -106,20 +158,25 @@ class ProductSearchContextProvider extends Component {
   }
 
   render() {
-    const props = processSearchContextProps(
-      this.props,
-      this.state,
-      DEFAULT_PAGE
-    )
     const {
+      nextTreePath,
       params,
-      map,
-      rest,
-      orderBy,
-      from,
-      to,
-      runtime: { page },
-    } = props
+      query: {
+        order: orderBy = SORT_OPTIONS[0].value,
+        page: pageProps,
+        map: mapProps,
+        rest = '',
+        priceRange,
+      },
+      runtime: { page: runtimePage },
+    } = this.props
+
+    const { variables: { maxItemsPerPage } } = this.state
+
+    const map = mapProps || createMap(params, rest)
+    const page = pageProps ? parseInt(pageProps) : DEFAULT_PAGE
+    const from = (page - 1) * maxItemsPerPage
+    const to = from + maxItemsPerPage - 1
 
     return (
       <Query
@@ -131,10 +188,12 @@ class ProductSearchContextProvider extends Component {
           map,
           rest,
           orderBy,
+          priceRange,
           from,
           to,
         }}
-        notifyOnNetworkStatusChange>
+        notifyOnNetworkStatusChange
+      >
         {searchQueryProps => {
           const { data, loading } = searchQueryProps
           const { search } = data || {}
@@ -146,7 +205,8 @@ class ProductSearchContextProvider extends Component {
                   ...search,
                 })
               }
-              loading={this.state.loading || loading}>
+              loading={this.state.loading || loading}
+            >
               <Helmet>
                 {params && (
                   <link
@@ -160,14 +220,22 @@ class ProductSearchContextProvider extends Component {
                 )}
               </Helmet>
               {React.cloneElement(this.props.children, {
-                loading: this.state.loading,
-                setContextVariables: this.handleContextVariables,
-                ...props,
+                ...this.props,
                 searchQuery: {
                   ...searchQueryProps,
                   ...search,
                 },
-                searchContext: page,
+                loading: this.state.loading,
+                setContextVariables: this.handleContextVariables,
+                searchContext: runtimePage,
+                pagesPath: nextTreePath,
+                map,
+                rest,
+                orderBy,
+                priceRange,
+                page,
+                from,
+                to,
               })}
             </DataLayerApolloWrapper>
           )